Consolidate mapping classes.
